requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
python-multipart
python-dotenv
